+++ tests/nxs_utils/test_detector.py
@@ -1,7 +1,14 @@
 # import pytest
 from numpy.testing import assert_array_equal
 
-from nexgen.nxs_utils import Axis, Detector, EigerDetector, TristanDetector
+from nexgen.nxs_utils import (
+    Axis,
+    Detector,
+    EigerDetector,
+    JungfrauDetector,
+    SinglaDetector,
+    TristanDetector,
+)
 from nexgen.utils import Point3D
 
 det_axes = [
@@ -11,6 +18,8 @@ det_axes = [
 
 test_eiger = EigerDetector("Eiger2 1M", (1028, 1062), "Si", 10000, -1)
 test_tristan = TristanDetector("Tristan 1M", (515, 2069))
+test_jungfrau = JungfrauDetector("Jungfrau 1M", (1066, 1030))
+test_singla = SinglaDetector("Singla 1M", (1062, 1028))
 
 
 def test_eiger_detector():
@@ -25,6 +34,18 @@ def test_tristan_detector():
     assert test_tristan.sensor_material == "Si"
     assert test_tristan.sensor_thickness == "0.5mm"
     assert test_tristan.mode == "events"
+
+
+def test_jungfrau_detector():
+    assert test_jungfrau.description == "Jungfrau 1M"
+    assert test_jungfrau.sensor_material == "Si"
+    assert test_jungfrau.sensor_thickness == "0.320mm"
+
+
+def test_singla_detector():
+    assert test_singla.description == "Singla 1M"
+    assert test_singla.sensor_thickness == "0.450mm"
+    assert test_singla.detector_type == "HPC"
 
 
 def test_detector_axes():